fix star rating precsion

+++ src/components/StarRating.jsx
@@ -23,7 +23,7 @@ export default function StarRating(props) {
         <div className={classes.root}>
             <Rating
                 name="hover-feedback"
-                value={parseInt(rating)}
+                value={parseFloat(rating)}
                 max={10}
                 precision={precision}
                 size="small"
